Moved all scripts into new src subfolder
