--- conf/script/build_system/cmd/hierarchy/find_root/__init__.py
@@ -1,2 +1,0 @@
-import build_system.cmd.hierarchy.find_root.cli
-from build_system.cmd.hierarchy.find_root.find_root import find_root, get_error_msg_root_not_found, VCS_DIR_NAME
