--- conf/script/build_system/cmd/hierarchy/find_root/__main__.py
@@ -1,8 +1,0 @@
-import utils.cli
-from build_system.cmd.hierarchy.find_root import find_root
-from build_system.cmd.hierarchy.find_root import cli
-
-if __name__ == '__main__':
-    utils.cli.init()
-    root = find_root(cli.get_error_formatted_msg_root_not_found)
-    print(root, end=str())
