--- conf/script/build_system/compiler/host/__init__.py
@@ -1,2 +1,0 @@
-from build_system.compiler.host.architecture import Architecture
-from build_system.compiler.host.os_family import OSFamily
