--- conf/script/build_system/compiler/host/get_info/location/msvc/__main__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python3
-
-import build_system.compiler.host.get_info.location.msvc.cli
-import utils.cli
-from build_system.compiler.host.get_info import location
-
-# Run as a script
-if __name__ == '__main__':
-    utils.cli.init()
-    location.msvc.cli.find()
