--- conf/script/build_system/compiler/host/get_info/version/gnu/__init__.py
@@ -1,2 +1,0 @@
-import build_system.compiler.host.get_info.version.gnu.cli
-from build_system.compiler.host.get_info.version.gnu.fetch_gnu_compiler_version import fetch
