--- conf/script/build_system/compiler/host/get_info/version/msvc/__init__.py
@@ -1,2 +1,0 @@
-import build_system.compiler.host.get_info.version.msvc.cli
-from build_system.compiler.host.get_info.version.msvc.fetch_msvc_version import fetch
