+++ conf/script/src/build_system/cmd/hierarchy/find_root/__init__.py
@@ -1,0 +1,1 @@
+import src.build_system.cmd.hierarchy.find_root.cli
