+++ conf/script/src/build_system/cmd/setup/build_type.py
