+++ conf/script/src/build_system/compiler/host/__init__.py
@@ -1,0 +1,1 @@
+from src.build_system.compiler.host.architecture import Architecture
