+++ conf/script/src/build_system/compiler/host/get_info/location/msvc/__main__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+
+from src import utils
+from build_system.compiler.host.get_info import location
+
+# Run as a script
+if __name__ == '__main__':
+    utils.cli.init()
+    location.msvc.cli.find()
