+++ conf/script/src/build_system/compiler/host/get_info/version/__init__.py
