+++ conf/script/src/build_system/compiler/host/get_info/version/gnu/__main__.py
