+++ conf/script/src/build_system/compiler/host/get_info/version/gnu/fetch_gnu_compiler_version.py
