+++ conf/script/src/build_system/compiler/host/get_info/version/msvc/__init__.py
@@ -1,0 +1,1 @@
+import src.build_system.compiler.host.get_info.version.msvc.cli
