+++ conf/script/src/build_system/compiler/host/get_info/version/msvc/cli.py
