+++ conf/script/src/build_system/compiler/host/get_info/version/msvc/fetch_msvc_version.py
