+++ conf/script/src/utils/auto_print.py
