+++ conf/script/src/utils/cmd_integrity.py
