+++ conf/script/src/utils/file_path_integrity.py
